// packages/beecli/sdk/beetrade-v2/models/CoreWatchlistTransportHttpCreateWatchListRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CoreWatchlistTransportHttpCreateWatchListRequest
 */
export interface CoreWatchlistTransportHttpCreateWatchListRequest {
    /**
     * 
     * @type {string}
     * @memberof CoreWatchlistTransportHttpCreateWatchListRequest
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof CoreWatchlistTransportHttpCreateWatchListRequest
     */
    name: string;
}

/**
 * Check if a given object implements the CoreWatchlistTransportHttpCreateWatchListRequest interface.
 */
export function instanceOfCoreWatchlistTransportHttpCreateWatchListRequest(value: object): value is CoreWatchlistTransportHttpCreateWatchListRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CoreWatchlistTransportHttpCreateWatchListRequestFromJSON(json: any): CoreWatchlistTransportHttpCreateWatchListRequest {
    return CoreWatchlistTransportHttpCreateWatchListRequestFromJSONTyped(json, false);
}

export function CoreWatchlistTransportHttpCreateWatchListRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CoreWatchlistTransportHttpCreateWatchListRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'description': json['description'] == null ? undefined : json['description'],
        'name': json['name'],
    };
}

export function CoreWatchlistTransportHttpCreateWatchListRequestToJSON(json: any): CoreWatchlistTransportHttpCreateWatchListRequest {
    return CoreWatchlistTransportHttpCreateWatchListRequestToJSONTyped(json, false);
}

export function CoreWatchlistTransportHttpCreateWatchListRequestToJSONTyped(value?: CoreWatchlistTransportHttpCreateWatchListRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'description': value['description'],
        'name': value['name'],
    };
}